macs2: improved typing of narrow_peaks output
Was incorrectly reduced to keep3 to make it compatible with
bedToBigBed. As is, narrow peaks is a non standard (BED5+5, although
they claim BED6+4) bed format, so it is not compatible with
bedToBigBed. The fix consists in expressing the actual type of narrow
peaks, and restricted to 5 cols only when actually computing the
bigBed (hence the new bed5_like and keep5 in Bed module).

As a note, this error was not detected before because bedToBigBed is
typed too generously. This should be fixed too.